All Glossary
finished up listing all lists with links but only allowing POST actions to happen if the user names match

+++ mysite/main/views.py
@@ -10,30 +10,34 @@ def index(response, id):
     print(response.POST)
  
     if response.method == "POST":
-        if response.POST.get("save"):
-            for item in ls.item_set.all():
-                if response.POST.get("c" + str(item.id)) == "clicked":
-                    item.complete = True
+        if response.user.username == ls.user.username:
+            print("sameuser")
+            if response.POST.get("save"):
+                for item in ls.item_set.all():
+                    if response.POST.get("c" + str(item.id)) == "clicked":
+                        item.complete = True
+                    else:
+                        item.complete = False
+        
+                    item.save()
+        
+            elif response.POST.get("newItem"):
+                txt = response.POST.get("new")
+        
+                if len(txt) > 2:
+                    ls.item_set.create(text=txt, complete=False)
                 else:
-                    item.complete = False
-    
-                item.save()
-    
-        elif response.POST.get("newItem"):
-            txt = response.POST.get("new")
-    
-            if len(txt) > 2:
-                ls.item_set.create(text=txt, complete=False)
-            else:
-                print("invalid")
+                    print("invalid")
 
-        elif response.POST.get("delItem"):
-            itemId = int(response.POST.get("delItem").split(",")[0])
-            print(itemId)
-            item = [item for item in ls.item_set.all() if itemId == item.id][0]
-            item.delete()
+            elif response.POST.get("delItem"):
+                itemId = int(response.POST.get("delItem").split(",")[0])
+                print(itemId)
+                item = [item for item in ls.item_set.all() if itemId == item.id][0]
+                item.delete()
 
-            return HttpResponseRedirect("/%i" %id)
+                return HttpResponseRedirect("/%i" %id)
+        else:
+            print("notsame")
  
  
     return render(response, "main/list.html", {"ls":ls})
@@ -51,6 +55,10 @@ def glossary(response):
 
  
     return render(response, "main/glossary.html", {})
+
+def glossaryall(response):
+    ToDos = ToDoList.objects.all()
+    return render(response, "main/glossaryall.html", {"ToDos":ToDos})
 
 def home(response):
     return render(response,"main/home.html",{})
